add configurational file

+++ ising.cpp
@@ -39,7 +39,7 @@ void print_summary(int *ising_matrix, long long int n) {
 
 }
 
-int get_energy_diff(const int *ising_matrix, int row, int col, int n) {
+int get_energy_diff(const int *ising_matrix, int row, int col, long long int n) {
 
   int up, down, left, right;
